pyproject.toml
[tool.uv]
cache-dir = "./.uv_cache"

[project]
name = "Studentprognose"
version = "0.1.0"
description = "Studentprognose, prediction of student numbers in the following (and upcoming) year(s)."
requires-python = ">=3.12"
dependencies = [
    "joblib>=1.5.2",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "statsmodels>=0.14.5",
    "xgboost>=3.0.5",
    "xlsxwriter>=3.2.9",
    "openpyxl>=3.1.2",
]

[tool.uv.build-backend]
module-root = ""
module-name = "scripts"

[project.scripts]
cumulative = "scripts.models.cumulative:main"
ratio = "scripts.models.ratio:main"
individual = "scripts.models.individual:main"
ensemble = "scripts.models.ensemble:main"
baseline = "scripts.models.baseline:main"
higher_years = "scripts.models.higher_years:main"
main = "scripts.main:main"
clear_output_dir = "scripts.standalone.clear_output_dir:main"
update_total_file = "scripts.standalone.update_total_file:main"
evaluate_results = "scripts.standalone.evaluate_results:main"
clear_cache = "scripts.utils.load_data:main"
postprocess = "scripts.utils.postprocess:main"
clustering = "scripts.utils.clustering:main"

# Add this section to guide package discovery
[tool.setuptools.packages.find]
include = ["scripts*"]  
